+++ app/renderer/types.js
@@ -52,6 +52,7 @@ const actionTypes = {
   stakingUnStake: apiActionTypes('staking_un_stake'),
   stakingSavePreferences: apiActionTypes('staking_save_preferences'),
   pendingToSendStakingExtrinsic: triggerActionTypes('pending_to_send_staking_extrinsic'),
+  stakingGetValidatorPreferences: apiActionTypes('staking_get_validator_preferences'),
 
   /* system apis */
   nodeJsonRpcSystem: apiActionTypes('node_jsonrpc_system'),